ingestion/requirements.txt
psycopg[binary]>=3.1.0
python-docx>=1.1.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
torch>=2.0.0
pgvector>=0.2.0
jupyter>=1.0.0
ipykernel>=6.0.0
pandas>=2.0.0
openpyxl>=3.1.0
matplotlib>=3.7.0
seaborn>=0.13.0
scikit-learn>=1.3.0
sqlalchemy>=2.0.0
umap-learn>=0.5.0
